package/qt6/qt6shadertools: new package
The Qt6::Gui module is needed when building the 'Qt Shader Baker'
(qsb) host tool.

Qsb is needed for converting the Qt6 Vulkan GLSL source code to
platform specific shader languages which is then later copied to
target.

For more details take a look at:

https://doc.qt.io/qt-6/qtshadertools-overview.html

and

https://doc.qt.io/qt-6/qshaderbaker.html

We need host-qt6base with Gui support when building host-qt6shadertools,
otherwise the build is skipped and no qsb host tool is generated:

  Skipping the build as the condition "TARGET Qt::Gui" is not met.

qt6shadertools fail to build if qsb is not available:

  Failed to find the host tool "Qt6::qsb".  It is part of the
  Qt6ShaderToolsTools package, but the package could not be found.  Make sure
  you have built and installed the host ShaderTools module, which will ensure
  the creation of the Qt6ShaderToolsTools package.