package/qt6/qt6declarative: new package
We need host-qt6base with Gui support when building host-qt6shadertools,
otherwise the build is skipped and no qsb host tool is generated.

qsb is required when building host-qt6declarative with Quick support,
which is necessary for the qmltime host tool to be built. qt6declarative
fails to build if qmltime is not available:

  Failed to find the host tool "Qt6::qmltime".  It is part of the Qt6QmlTools
  package, but the package did not contain the tool.  Make sure that the host
  module Qml was built with all features enabled (no explicitly disabled
  tools).

Gui support is also required when building Svg support and thereby the
svgtoqml host tool. qt6declarative will also fail to build if svgtoqmlis not
available:

  Failed to find the host tool "Qt6::svgtoqml".  It is part of the
  Qt6QuickTools package, but the package could not be found.  Make sure you
  have built and installed the host Quick module, which will ensure the
  creation of the Qt6QuickTools package.

We need host-qt6base with Testlib support when building host-qt6declarative
with QuickTest support. QuickTest support is further required for building the
qmltestrunner host tool. qt6declarative will fail to build if qmltestrunner is
not available:

  Failed to find the host tool "Qt6::qmltestrunner".  It is part of the
  Qt6QmlTools package, but the package did not contain the tool.  Make sure
  that the host module Qml was built with all features enabled (no explicitly
  disabled tools).

We add the patch to avoid checking for FEATURE_ssl when FEATURE_network
is not available. FEATURE_ssl is defined by the Qt Network module. So
we will get the following build failure if the patch is not applied and
FEATURE_network=ON:

  Attempting to evaluate feature ssl but its definition is missing.  Either
  the feature does not exist or a dependency to the module that defines it is
  missing