package/qt6/qt6base: add blind option to enable GUI support on host
We need host-qt6base with Gui support when building host-qt6shadertools,
otherwise the build is skipped and no qsb host tool is generated.
qt6shadertools fail to build if qsb is not available. qsb is also required
when building host-qt6declarative with Quick support, which is necessary for
the qmltime host tool to be built. qt6declarative fails to build if qmltime is
not available. Gui support is also required when building Svg support and
thereby the svgtoqml host tool. qt6declarative will also fail to build if
svgtoqml is not available.